+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2025-09-13
+
+### Fixed
+
+- `People` related resource properties have been fixed and are now collection types.
+
+### Changed
+
+- Move `AddPlanningCenterOAuth` extension methods to the `DependencyInjection` namespace for simplicity.
+- Change `AddPlanningCenterOAuth` extension methods to use the `PlanningCenter:ClientId` and `PlanningCenter:ClientSecret` configuration sections by default.
+
 ## [1.2.0] - 2025-08-27
 
 ### Added